fix: refactor away prepare_chat_input and improve tool grammar apply control flow
